stavio JavuAlert,sredio style

+++ Expert.WebShop.Vjezbe/models/BuyItems.cs
@@ -14,11 +14,11 @@
 
         public int NumberOfItems { get; set; }
 
-        public float PricePerProduct { get; set; }
+        public decimal PricePerProduct { get; set; }
 
         public int ProductDiscount { get; set; }
 
-        public float TotalAmmount { get; set; }
+        public decimal TotalAmmount { get; set; }
 
         public bool IsPayed { get; set; }
